requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# Supabase
supabase==2.10.0

# Configuration
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Security / Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# LLM Providers
openai==1.57.4
anthropic==0.40.0
google-generativeai==0.8.3

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
tenacity==9.0.0
structlog==24.4.0
slowapi==0.1.9
youtube-search-python==1.6.6

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
